+++ projects/ccmanager/ccmanager.h
@@ -13,5 +13,4 @@
     void writeConfigFile(configFormat* config, char* filepath);
     char* parseFormatToString(configFormat* config);
     void appendConfigFile(configFormat* config, char* filepath);
-    void capCheck(configFormat* config, size_t* size);
 #endif