+++ projects/cfm/cfm.h
@@ -11,7 +11,7 @@
     char* getListedDirectories(char* directory);
     char* getRootFilePath();
     void* findFile(char* name, char* root, char* type);
-    char* ParseFile(FILE* fp, size_t* line);
+    char** ParseFile(FILE* fp, size_t* line);
     WIN32_FIND_DATA* listFiles(char* key, char* path);
     int dirExists(const char* path);
     TCHAR* getCurrentDirectory();